requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
tqdm>=4.65.0
pyyaml>=6.0
scipy>=1.11.0
scikit-learn>=1.3.0
ultralytics>=8.0.0
pycocotools>=2.0.6
pytest>=7.4.0
pytest-cov>=4.1.0
flake8>=6.1.0
